refactor: extract shared TripCard, FuelCard, ExpenseCard, ReservationCard components (#18)
